fix(admin): 화면 크기 변경 시 편집 상태 유실 방지
