+++ frontend/src/pages/AdminPage/tabs/PhotoEditTab/PhotoEditTabDesktop.tsx
@@ -1,0 +1,145 @@
+import { useRef } from 'react';
+import Button from '@/components/common/Button/Button';
+import { ADMIN_EVENT } from '@/constants/eventName';
+import { MAX_FILE_COUNT } from '@/constants/uploadLimit';
+import useMixpanelTrack from '@/hooks/Mixpanel/useMixpanelTrack';
+import { ContentSection } from '@/pages/AdminPage/components/ContentSection/ContentSection';
+import { FeedImageGrid } from './components/FeedImageGrid/FeedImageGrid';
+import { useDragSort } from './hooks/useDragSort';
+import { FeedItem } from './types';
+import * as Styled from './PhotoEditTab.styles';
+
+interface PhotoEditTabDesktopProps {
+  feedItems: FeedItem[];
+  feedItemsRef: React.MutableRefObject<FeedItem[]>;
+  setFeedItems: React.Dispatch<React.SetStateAction<FeedItem[]>>;
+  isLoading: boolean;
+  pendingChanges: boolean;
+  addFiles: (files: File[]) => void;
+  deleteImage: (index: number) => void;
+  clearAll: () => void;
+  retryItem: (index: number) => void;
+  save: () => void;
+}
+
+const PhotoEditTabDesktop = ({
+  feedItems,
+  feedItemsRef,
+  setFeedItems,
+  isLoading,
+  pendingChanges,
+  addFiles,
+  deleteImage,
+  clearAll,
+  retryItem,
+  save,
+}: PhotoEditTabDesktopProps) => {
+  const trackEvent = useMixpanelTrack();
+  const inputRef = useRef<HTMLInputElement>(null);
+  const isFull = feedItems.length >= MAX_FILE_COUNT;
+
+  const { gridRef, dragIndex, dropPosition, handleMouseDown } = useDragSort({
+    disabled: isLoading,
+    onReorder: setFeedItems,
+    feedItemsRef,
+  });
+
+  const handleAddClick = () => {
+    if (isLoading || isFull) return;
+    trackEvent(ADMIN_EVENT.IMAGE_UPLOAD_BUTTON_CLICKED);
+    inputRef.current?.click();
+  };
+
+  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    if (!e.target.files?.length) return;
+    addFiles(Array.from(e.target.files));
+    e.target.value = '';
+  };
+
+  const handleClearAll = () => {
+    if (isLoading || !window.confirm('모든 사진을 삭제하시겠어요?')) return;
+    clearAll();
+  };
+
+  return (
+    <Styled.Container>
+      <ContentSection>
+        <ContentSection.Header
+          title='활동 사진'
+          action={
+            <Button
+              width={'150px'}
+              animated
+              onClick={save}
+              disabled={isLoading || !pendingChanges}
+            >
+              {isLoading && <Styled.ButtonSpinner />}
+              {isLoading ? '저장 중...' : '저장하기'}
+            </Button>
+          }
+        />
+
+        <ContentSection.Body>
+          <input
+            ref={inputRef}
+            type='file'
+            accept='image/*'
+            multiple
+            hidden
+            onChange={handleFileChange}
+          />
+
+          {feedItems.length > 0 && (
+            <Styled.GridHeader>
+              <Styled.AddButton
+                onClick={handleAddClick}
+                disabled={isLoading || isFull}
+              >
+                + 이미지 추가
+              </Styled.AddButton>
+              <Styled.ClearAllButton
+                onClick={handleClearAll}
+                disabled={isLoading}
+              >
+                전체 삭제
+              </Styled.ClearAllButton>
+            </Styled.GridHeader>
+          )}
+
+          <Styled.GridWrapper $uploading={isLoading}>
+            {isLoading && (
+              <Styled.UploadOverlay>
+                <Styled.OverlaySpinner />
+                <span>사진을 업로드하고 있어요</span>
+              </Styled.UploadOverlay>
+            )}
+            {feedItems.length === 0 ? (
+              <Styled.EmptyState onClick={handleAddClick}>
+                <span>+</span>
+                <span>사진을 추가해보세요</span>
+                <span>최대 {MAX_FILE_COUNT}장</span>
+              </Styled.EmptyState>
+            ) : (
+              <FeedImageGrid
+                feedItems={feedItems}
+                gridRef={gridRef}
+                dragIndex={dragIndex}
+                dropPosition={dropPosition}
+                isLoading={isLoading}
+                columns={4}
+                onMouseDown={handleMouseDown}
+                onDelete={(index) => {
+                  trackEvent(ADMIN_EVENT.IMAGE_DELETE_BUTTON_CLICKED);
+                  deleteImage(index);
+                }}
+                onRetry={retryItem}
+              />
+            )}
+          </Styled.GridWrapper>
+        </ContentSection.Body>
+      </ContentSection>
+    </Styled.Container>
+  );
+};
+
+export default PhotoEditTabDesktop;
